requirements.txt
certifi==2020.6.20
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.5.3
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
